requirements.txt
streamlit==1.48.1
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.2
gspread
google-auth
plotly
